fix messages decrypting
