+++ FlowCrypt/Functionality/Mail Provider/Message Provider/MessageService.swift
@@ -86,18 +86,19 @@ final class MessageService {
     }
 
     // MARK: - Message processing
-    func getAndProcess(
-        identifier: Identifier,
-        folder: String,
+    func fetchMessage(identifier: Identifier, folder: String) async throws -> Message {
+        return try await messageProvider.fetchMessage(
+            id: identifier,
+            folder: folder
+        )
+    }
+
+    func process(
+        message: Message,
         onlyLocalKeys: Bool,
         userEmail: String,
         isUsingKeyManager: Bool
     ) async throws -> ProcessedMessage {
-        let message = try await messageProvider.fetchMessage(
-            id: identifier,
-            folder: folder
-        )
-
         guard message.isPgp else {
             return ProcessedMessage(message: message)
         }
@@ -182,6 +183,9 @@ final class MessageService {
             isMime: message.raw != nil,
             verificationPubKeys: verificationPubKeys
         )
+
+        print("decrypt")
+        print(decrypted.text)
 
         return try await process(
             message: message,
